Various compile fixes and code adjustments for kernel 6.6. Warning: Like kernel 6.6 we now use 64 Bit integers for capabilities instead of an array of two 32 Bit, this requires new list on-disk versions for CAP FD and USER attributes. Existing lists get converted as usual, but after they got saved for the first time there is no way back to read them with older kernels.

+++ drivers/tty/sysrq.c
@@ -214,7 +214,7 @@ static const struct sysrq_key_op sysrq_mountro_op = {
 };
 
 #ifdef CONFIG_RSBAC_SOFTMODE_SYSRQ
-static void sysrq_handle_rsbac_softmode(int key) {
+static void sysrq_handle_rsbac_softmode(u8 key) {
 	if (rsbac_softmode) {
 		rsbac_printk(KERN_WARNING "Soft mode disabled via SysRq!\n");
 		rsbac_softmode = 0;
